Don't clutter project repo with .*.swp files
I ran into an issue while developing salt states wherein salt would try to copy
a folder with files I had open in vim. It would error out on the binary .swp
files. This patch is more of a workaorund than a fix, but the true issue lies
in salt.